pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "fastapi>=0.128.0",
    "gunicorn>=23.0.0",
    "passlib==1.7.4",
    "pydantic-settings>=2.12.0",
    "python-decouple>=3.8",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "types-passlib>=1.7.7.20250602",
]
test = [
    "faker>=40.1.2",
    "pytest>=9.0.2",
]

[tool.black]
line-length = 79
